+++ mlflow/experiments.py
@@ -3,7 +3,7 @@ import os
 import click
 
 import mlflow
-from mlflow.data import is_uri
+from mlflow.utils.data_utils import is_uri
 from mlflow.entities import ViewType
 from mlflow.tracking import _get_store, fluent
 from mlflow.utils.string_utils import _create_table
