+++ mlflow/store/artifact/s3_artifact_repo.py
@@ -6,7 +6,6 @@ from mimetypes import guess_type
 import posixpath
 import urllib.parse
 
-from mlflow import data
 from mlflow.entities import FileInfo
 from mlflow.environment_variables import (
     MLFLOW_S3_UPLOAD_EXTRA_ARGS,
@@ -15,6 +14,7 @@ from mlflow.environment_variables import (
 )
 from mlflow.exceptions import MlflowException
 from mlflow.store.artifact.artifact_repo import ArtifactRepository
+from mlflow.utils import data_utils
 from mlflow.utils.file_utils import relative_path_to_artifact_path
 
 _MAX_CACHE_SECONDS = 300
@@ -147,7 +147,7 @@ class S3ArtifactRepository(ArtifactRepository):
         s3_client.upload_file(Filename=local_file, Bucket=bucket, Key=key, ExtraArgs=extra_args)
 
     def log_artifact(self, local_file, artifact_path=None):
-        (bucket, dest_path) = data.parse_s3_uri(self.artifact_uri)
+        (bucket, dest_path) = data_utils.parse_s3_uri(self.artifact_uri)
         if artifact_path:
             dest_path = posixpath.join(dest_path, artifact_path)
         dest_path = posixpath.join(dest_path, os.path.basename(local_file))
@@ -156,7 +156,7 @@ class S3ArtifactRepository(ArtifactRepository):
         )
 
     def log_artifacts(self, local_dir, artifact_path=None):
-        (bucket, dest_path) = data.parse_s3_uri(self.artifact_uri)
+        (bucket, dest_path) = data_utils.parse_s3_uri(self.artifact_uri)
         if artifact_path:
             dest_path = posixpath.join(dest_path, artifact_path)
         s3_client = self._get_s3_client()
@@ -176,7 +176,7 @@ class S3ArtifactRepository(ArtifactRepository):
                 )
 
     def list_artifacts(self, path=None):
-        (bucket, artifact_path) = data.parse_s3_uri(self.artifact_uri)
+        (bucket, artifact_path) = data_utils.parse_s3_uri(self.artifact_uri)
         dest_path = artifact_path
         if path:
             dest_path = posixpath.join(dest_path, path)
@@ -217,13 +217,13 @@ class S3ArtifactRepository(ArtifactRepository):
             )
 
     def _download_file(self, remote_file_path, local_path):
-        (bucket, s3_root_path) = data.parse_s3_uri(self.artifact_uri)
+        (bucket, s3_root_path) = data_utils.parse_s3_uri(self.artifact_uri)
         s3_full_path = posixpath.join(s3_root_path, remote_file_path)
         s3_client = self._get_s3_client()
         s3_client.download_file(bucket, s3_full_path, local_path)
 
     def delete_artifacts(self, artifact_path=None):
-        (bucket, dest_path) = data.parse_s3_uri(self.artifact_uri)
+        (bucket, dest_path) = data_utils.parse_s3_uri(self.artifact_uri)
         if artifact_path:
             dest_path = posixpath.join(dest_path, artifact_path)
